rag_ingestion_service/requirements.txt
google-cloud-storage==2.10.0
google-cloud-aiplatform==1.38.1
qdrant-client==1.7.0
requests==2.31.0
python-dotenv==1.0.0
pydantic==2.5.0
numpy==1.24.3
tiktoken==0.5.2
langchain==0.1.0
langchain-community==0.0.10
langchain-google-vertexai==0.0.6
sentence-transformers==2.2.2
markdown==3.5.1
beautifulsoup4==4.12.2
flask==3.0.0
vertexai==1.38.1
